move generate command to its own file
